requirements.txt
flask==3.1.0
flask_wtf==1.2.2
python-dotenv==1.0.1
wtforms==3.2.1
bleach==6.2.0
flask_login==0.6.3
flask_socketio==5.5.1
cloudinary==1.42.1
sqlalchemy==2.0.37
werkzeug==3.1.3
flask_sqlalchemy==3.1.1